Split channel tests
