+++ saleor/graphql/channel/tests/mutations/test_channel_update.py
@@ -7,12 +7,12 @@ from django.utils.functional import SimpleLazyObject
 from django.utils.text import slugify
 from freezegun import freeze_time
 
-from ....channel.error_codes import ChannelErrorCode
-from ....core.utils.json_serializer import CustomJsonEncoder
-from ....webhook.event_types import WebhookEventAsyncType
-from ....webhook.payloads import generate_meta, generate_requestor
-from ...tests.utils import assert_no_permission, get_graphql_content
-from ..enums import AllocationStrategyEnum
+from .....channel.error_codes import ChannelErrorCode
+from .....core.utils.json_serializer import CustomJsonEncoder
+from .....webhook.event_types import WebhookEventAsyncType
+from .....webhook.payloads import generate_meta, generate_requestor
+from ....tests.utils import assert_no_permission, get_graphql_content
+from ...enums import AllocationStrategyEnum
 
 CHANNEL_UPDATE_MUTATION = """
     mutation UpdateChannel($id: ID!,$input: ChannelUpdateInput!){
